Move modules out of setup
