+++ roms_tools/mask.py
