+++ roms_tools/setup/boundary_forcing.py
@@ -5,11 +5,14 @@ import logging
 import importlib.metadata
 from typing import Dict, Union, List
 from dataclasses import dataclass, field
-from roms_tools.setup.grid import Grid
-from roms_tools.setup.regrid import LateralRegrid, VerticalRegrid
 from datetime import datetime
+import matplotlib.pyplot as plt
+from pathlib import Path
+from roms_tools.grid import Grid
+from roms_tools.regrid import LateralRegrid, VerticalRegrid
+from roms_tools.vertical_coordinate import compute_depth
+from roms_tools.plot import _section_plot, _line_plot
 from roms_tools.setup.datasets import GLORYSDataset, CESMBGCDataset
-from roms_tools.setup.vertical_coordinate import compute_depth
 from roms_tools.setup.utils import (
     get_variable_metadata,
     group_dataset,
@@ -28,9 +31,6 @@ from roms_tools.setup.utils import (
     _to_yaml,
     _from_yaml,
 )
-from roms_tools.setup.plot import _section_plot, _line_plot
-import matplotlib.pyplot as plt
-from pathlib import Path
 
 
 @dataclass(frozen=True, kw_only=True)
